requirements-vercel.txt
# Vercel: no Chroma (use embeddings.json). Keeps bundle under 250MB.
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
numpy>=1.24.0
httpx>=0.26.0
python-dotenv>=1.0.0
PyYAML>=6.0
